+++ tests/test_client.py
@@ -39,7 +39,7 @@ from pilosa import TimeQuantum, CacheType
 from pilosa.client import Client, URI, Cluster, _QueryRequest, \
     decode_field_meta_options, _ImportRequest, _Node
 from pilosa.exceptions import PilosaURIError, PilosaError
-from pilosa.imports import Bit
+from pilosa.imports import Columns
 
 logger = logging.getLogger(__name__)
 
@@ -59,9 +59,9 @@ class ClientTestCase(unittest.TestCase):
         # create with invalid type
         self.assertRaises(PilosaError, Client, 15000)
 
-    def test_decode_frame_meta_options(self):
-        frame_info = {}
-        options = decode_field_meta_options(frame_info)
+    def test_decode_field_meta_options(self):
+        field_info = {}
+        options = decode_field_meta_options(field_info)
         target = {
             "cache_size": 50000,
             "cache_type": CacheType.DEFAULT,
@@ -220,7 +220,7 @@ class ClusterTestCase(unittest.TestCase):
 class QueryRequestTestCase(unittest.TestCase):
 
     def test_serialize(self):
-        qr = _QueryRequest("Bitmap(field='foo', id=1)", columns=True)
+        qr = _QueryRequest("Bitmap(field='foo', id=1)", column_attrs=True)
         bin = qr.to_protobuf(False)  # do not return a bytearray
         self.assertIsNotNone(bin)
         qr = internal.QueryRequest()
@@ -232,13 +232,13 @@ class QueryRequestTestCase(unittest.TestCase):
 class ImportRequestTestCase(unittest.TestCase):
 
     def test_serialize(self):
-        ir = _ImportRequest("foo", "bar", 0, [Bit(row_id=1, column_id=2, timestamp=3)])
+        ir = _ImportRequest("foo", "bar", 0, [Columns(row_id=1, column_id=2, timestamp=3)])
         bin = ir.to_protobuf(False)
         self.assertIsNotNone(bin)
         ir = internal.ImportRequest()
         ir.ParseFromString(bin)
         self.assertEquals("foo", ir.Index)
-        self.assertEquals("bar", ir.Frame)
+        self.assertEquals("bar", ir.Field)
         self.assertEquals([1], ir.RowIDs)
         self.assertEquals([2], ir.ColumnIDs)
         self.assertEquals([3], ir.Timestamps)