+++ tests/test_orm.py
@@ -39,9 +39,9 @@ from pilosa.orm import Schema
 
 schema = Schema()
 sampleIndex = schema.index("sample-db")
-sampleFrame = sampleIndex.field("sample-field")
+sampleField = sampleIndex.field("sample-field")
 projectIndex = schema.index("project-db")
-collabFrame = projectIndex.field("collaboration")
+collabField = projectIndex.field("collaboration")
 
 
 class SchemaTestCase(unittest.TestCase):
@@ -102,10 +102,10 @@ class IndexTestCase(unittest.TestCase):
             q.serialize())
 
     def test_union(self):
-        b1 = sampleFrame.bitmap(10)
-        b2 = sampleFrame.bitmap(20)
-        b3 = sampleFrame.bitmap(42)
-        b4 = collabFrame.bitmap(2)
+        b1 = sampleField.row(10)
+        b2 = sampleField.row(20)
+        b3 = sampleField.row(42)
+        b4 = collabField.row(2)
 
         q1 = sampleIndex.union(b1, b2)
         self.assertEquals(
@@ -123,10 +123,10 @@ class IndexTestCase(unittest.TestCase):
             q3.serialize())
 
     def test_intersect(self):
-        b1 = sampleFrame.bitmap(10)
-        b2 = sampleFrame.bitmap(20)
-        b3 = sampleFrame.bitmap(42)
-        b4 = collabFrame.bitmap(2)
+        b1 = sampleField.row(10)
+        b2 = sampleField.row(20)
+        b3 = sampleField.row(42)
+        b4 = collabField.row(2)
 
         q1 = sampleIndex.intersect(b1, b2)
         self.assertEquals(
@@ -144,10 +144,10 @@ class IndexTestCase(unittest.TestCase):
             q3.serialize())
 
     def test_difference(self):
-        b1 = sampleFrame.bitmap(10)
-        b2 = sampleFrame.bitmap(20)
-        b3 = sampleFrame.bitmap(42)
-        b4 = collabFrame.bitmap(2)
+        b1 = sampleField.row(10)
+        b2 = sampleField.row(20)
+        b3 = sampleField.row(42)
+        b4 = collabField.row(2)
 
         q1 = sampleIndex.difference(b1, b2)
         self.assertEquals(
@@ -165,8 +165,8 @@ class IndexTestCase(unittest.TestCase):
             q3.serialize())
 
     def test_xor(self):
-        b1 = sampleFrame.bitmap(10)
-        b2 = sampleFrame.bitmap(20)
+        b1 = sampleField.row(10)
+        b2 = sampleField.row(20)
         q1 = sampleIndex.xor(b1, b2)
 
         self.assertEquals(
@@ -178,7 +178,7 @@ class IndexTestCase(unittest.TestCase):
         self.assertEquals("Union()", q.serialize())
 
     def test_union1(self):
-        q = sampleIndex.union(sampleFrame.bitmap(10))
+        q = sampleIndex.union(sampleField.row(10))
         self.assertEquals("Union(Bitmap(row=10, field='sample-field'))", q.serialize())
 
     def test_intersect_invalid_bitmap_count_fails(self):
@@ -188,10 +188,10 @@ class IndexTestCase(unittest.TestCase):
         self.assertRaises(PilosaError, projectIndex.difference)
 
     def test_xor_invalid_bitmap_count_fails(self):
-        self.assertRaises(PilosaError, projectIndex.xor, sampleFrame.bitmap(10))
+        self.assertRaises(PilosaError, projectIndex.xor, sampleField.row(10))
 
     def test_count(self):
-        b = collabFrame.bitmap(42)
+        b = collabField.row(42)
         q = projectIndex.count(b)
         self.assertEquals(
             "Count(Bitmap(row=42, field='collaboration'))",
@@ -215,7 +215,7 @@ class IndexTestCase(unittest.TestCase):
         self.assertRaises(PilosaError, projectIndex.set_column_attrs, 5, attrs_map)
 
 
-class FrameTestCase(unittest.TestCase):
+class FieldTestCase(unittest.TestCase):
 
     def test_create_field(self):
         db = Index("foo")
@@ -225,94 +225,94 @@ class FrameTestCase(unittest.TestCase):
         self.assertEqual(TimeQuantum.NONE, field.time_quantum)
 
     def test_same_equals(self):
-        self.assertEqual(sampleFrame, sampleFrame)
+        self.assertEqual(sampleField, sampleField)
 
     def test_other_class_not_equals(self):
         schema = Schema()
-        self.assertNotEqual(sampleFrame, schema)
+        self.assertNotEqual(sampleField, schema)
 
     def test_bitmap(self):
-        qry1 = collabFrame.bitmap(5)
+        qry1 = collabField.row(5)
         self.assertEquals(
             "Bitmap(row=5, field='collaboration')",
             qry1.serialize())
 
-        qry2 = collabFrame.bitmap(10)
+        qry2 = collabField.row(10)
         self.assertEquals(
             "Bitmap(row=10, field='collaboration')",
             qry2.serialize())
 
-        qry3 = collabFrame.bitmap("b7feb014-8ea7-49a8-9cd8-19709161ab63")
+        qry3 = collabField.row("b7feb014-8ea7-49a8-9cd8-19709161ab63")
         self.assertEquals(
             "Bitmap(row='b7feb014-8ea7-49a8-9cd8-19709161ab63', field='collaboration')",
             qry3.serialize())
 
     def test_bitmap_with_invalid_id_type(self):
-        self.assertRaises(ValidationError, sampleFrame.bitmap, {})
+        self.assertRaises(ValidationError, sampleField.row, {})
 
     def test_setbit(self):
-        qry1 = collabFrame.setbit(5, 10)
+        qry1 = collabField.setbit(5, 10)
         self.assertEquals(
             "SetBit(row=5, field='collaboration', col=10)",
             qry1.serialize())
 
-        qry2 = collabFrame.setbit(10, 20)
+        qry2 = collabField.setbit(10, 20)
         self.assertEquals(
             "SetBit(row=10, field='collaboration', col=20)",
 
             qry2.serialize())
 
-        qry3 = collabFrame.setbit("b7feb014-8ea7-49a8-9cd8-19709161ab63", "some_id")
+        qry3 = collabField.setbit("b7feb014-8ea7-49a8-9cd8-19709161ab63", "some_id")
         self.assertEquals(
             "SetBit(row='b7feb014-8ea7-49a8-9cd8-19709161ab63', field='collaboration', col='some_id')",
             qry3.serialize())
 
     def test_setbit_with_invalid_id_type(self):
-        self.assertRaises(ValidationError, sampleFrame.setbit, {}, 1)
-        self.assertRaises(ValidationError, sampleFrame.setbit, 1, {})
-        self.assertRaises(ValidationError, sampleFrame.setbit, 1, "zero")
+        self.assertRaises(ValidationError, sampleField.setbit, {}, 1)
+        self.assertRaises(ValidationError, sampleField.setbit, 1, {})
+        self.assertRaises(ValidationError, sampleField.setbit, 1, "zero")
 
     def test_setbit_with_timestamp(self):
         timestamp = datetime(2017, 4, 24, 12, 14)
-        qry = collabFrame.setbit(10, 20, timestamp)
+        qry = collabField.setbit(10, 20, timestamp)
         self.assertEquals(
             "SetBit(row=10, field='collaboration', col=20, timestamp='2017-04-24T12:14')",
             qry.serialize()
         )
 
     def test_clearbit(self):
-        qry1 = collabFrame.clearbit(5, 10)
+        qry1 = collabField.clearbit(5, 10)
         self.assertEquals(
             "ClearBit(row=5, field='collaboration', col=10)",
             qry1.serialize())
 
-        qry2 = collabFrame.clearbit(10, 20)
+        qry2 = collabField.clearbit(10, 20)
         self.assertEquals(
             "ClearBit(row=10, field='collaboration', col=20)",
             qry2.serialize())
 
-        qry3 = collabFrame.clearbit("b7feb014-8ea7-49a8-9cd8-19709161ab63", "some_id")
+        qry3 = collabField.clearbit("b7feb014-8ea7-49a8-9cd8-19709161ab63", "some_id")
         self.assertEquals(
             "ClearBit(row='b7feb014-8ea7-49a8-9cd8-19709161ab63', field='collaboration', col='some_id')",
             qry3.serialize())
 
     def test_clearbit_with_invalid_id_type(self):
-        self.assertRaises(ValidationError, sampleFrame.clearbit, {}, 1)
-        self.assertRaises(ValidationError, sampleFrame.clearbit, 1, {})
-        self.assertRaises(ValidationError, sampleFrame.clearbit, 1, "zero")
+        self.assertRaises(ValidationError, sampleField.clearbit, {}, 1)
+        self.assertRaises(ValidationError, sampleField.clearbit, 1, {})
+        self.assertRaises(ValidationError, sampleField.clearbit, 1, "zero")
 
     def test_topn(self):
-        q1 = collabFrame.topn(27)
+        q1 = collabField.topn(27)
         self.assertEquals(
             "TopN(field='collaboration', n=27)",
             q1.serialize())
 
-        q2 = collabFrame.topn(10, collabFrame.bitmap(3))
+        q2 = collabField.topn(10, collabField.row(3))
         self.assertEquals(
             u"TopN(Bitmap(row=3, field='collaboration'), field='collaboration', n=10)",
             q2.serialize())
 
-        q3 = sampleFrame.topn(12, collabFrame.bitmap(7), "category", 80, 81)
+        q3 = sampleField.topn(12, collabField.row(7), "category", 80, 81)
         self.assertEquals(
             "TopN(Bitmap(row=7, field='collaboration'), field='sample-field', n=12, field='category', filters=[80,81])",
             q3.serialize())
@@ -321,12 +321,12 @@ class FrameTestCase(unittest.TestCase):
         start = datetime(1970, 1, 1, 0, 0)
         end = datetime(2000, 2, 2, 3, 4)
 
-        q1 = collabFrame.range(10, start, end)
+        q1 = collabField.range(10, start, end)
         self.assertEquals(
             "Range(row=10, field='collaboration', start='1970-01-01T00:00', end='2000-02-02T03:04')",
             q1.serialize())
 
-        q3 = sampleFrame.range("b7feb014-8ea7-49a8-9cd8-19709161ab63", start, end)
+        q3 = sampleField.range("b7feb014-8ea7-49a8-9cd8-19709161ab63", start, end)
         self.assertEquals(
             u"Range(row='b7feb014-8ea7-49a8-9cd8-19709161ab63', field='sample-field', start='1970-01-01T00:00', end='2000-02-02T03:04')",
             q3.serialize())
@@ -336,71 +336,71 @@ class FrameTestCase(unittest.TestCase):
             "quote": '''"Don't worry, be happy"''',
             "active": True
         }
-        q = collabFrame.set_row_attrs(5, attrs_map)
+        q = collabField.set_row_attrs(5, attrs_map)
         self.assertEquals(
             "SetRowAttrs(row=5, field='collaboration', active=true, quote=\"\\\"Don't worry, be happy\\\"\")",
             q.serialize())
 
     def test_field_lt(self):
-        q = collabFrame.lt(10)
+        q = collabField.lt(10)
         self.assertEquals(
             "Range(collaboration < 10)",
             q.serialize())
 
     def test_field_lte(self):
-        q = collabFrame.lte(10)
+        q = collabField.lte(10)
         self.assertEquals(
             "Range(collaboration <= 10)",
             q.serialize())
 
     def test_field_gt(self):
-        q = collabFrame.gt(10)
+        q = collabField.gt(10)
         self.assertEquals(
             "Range(collaboration > 10)",
             q.serialize())
 
     def test_field_gte(self):
-        q = collabFrame.gte(10)
+        q = collabField.gte(10)
         self.assertEquals(
             "Range(collaboration >= 10)",
             q.serialize())
 
     def test_field_equals(self):
-        q = collabFrame.equals(10)
+        q = collabField.equals(10)
         self.assertEquals(
             "Range(collaboration == 10)",
             q.serialize())
 
     def test_field_not_equals(self):
-        q = collabFrame.not_equals(10)
+        q = collabField.not_equals(10)
         self.assertEquals(
             "Range(collaboration != 10)",
             q.serialize())
 
     def test_field_not_null(self):
-        q = collabFrame.not_null()
+        q = collabField.not_null()
         self.assertEquals(
             "Range(collaboration != null)",
             q.serialize())
 
     def test_field_between(self):
-        q = collabFrame.between(10, 20)
+        q = collabField.between(10, 20)
         self.assertEquals(
             "Range(collaboration >< [10,20])",
             q.serialize())
 
     def test_field_sum(self):
-        q = collabFrame.sum(collabFrame.bitmap(10))
+        q = collabField.sum(collabField.row(10))
         self.assertEquals(
             "Sum(Bitmap(row=10, field='collaboration'), field='collaboration')",
             q.serialize())
-        q = collabFrame.sum()
+        q = collabField.sum()
         self.assertEquals(
             "Sum(field='collaboration')",
             q.serialize())
 
     def test_field_set_value(self):
-        q = collabFrame.set_value(10, 20)
+        q = collabField.setvalue(10, 20)
         self.assertEquals(
             "SetValue(col=10, collaboration=20)",
             q.serialize())